+++ scheduler.h
@@ -21,5 +21,5 @@ void print_process_table(struct Process p[],int n);
 void print_process_table_priority(struct Process p[],int n);
 void priority_np(struct Process p[],int n);
 void fcfs(struct Process p[],int n);
-void rr(struct Process p[],int n,int tq);
+void rr(struct Process p[],int n,float tq);
 #endif
